requirement.txt
streamlit>=1.52.0
pandas>=2.0
matplotlib>=3.8
python-dotenv>=1.0
langchain>=1.2.0
langchain-experimental>=0.3.0
langchain-groq>=1.1.0
groq>=0.37.0
tabulate>=0.9.0